AI_Resume_Ranking/requirements.txt

streamlit
langchain
faiss-cpu
python-docx
pdfminer.six
ollama
tiktoken
pandas
